# public/openapi/moim-api.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "MOIM API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://{env}.moim.id/api/v1",
            "variables": {
                "env": {
                    "enum": [
                        "dev",
                        "www"
                    ],
                    "default": "dev"
                }
            }
        }
    ],
...
    "components": {
        "schemas": {
            "ApiResponse": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer"
                    },
                    "success": {
                        "type": "boolean"
                    },
                    "messages": {
                        "type": "string"
                    }
                }
            }
        }
    },
    "paths": {
...
        "/post": {
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Untuk menambahkan posting",
                "description": "",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User Id",
                        "required": true,
                        "schema": {}
                    },
                    {
                        "name": "app-token",
                        "in": "header",
                        "description": "",
                        "required": false,
                        "schema": {}
                    }
                ],
                "requestBody": {
                    "description": "",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    },
                                    "moim_id": {
                                        "type": "string"
                                    },
                                    "attachment": {
                                        "type": "array",
                                        "items": {
                                            "type": "object",
                                            "properties": {
                                                "content": {
                                                    "type": "string",
                                                    "format": "#base64"
                                                },
                                                "extention": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "appliation/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "type": "object",
                                                    "properties": {
                                                        "moim_id": {
                                                            "type": "string"
                                                        },
                                                        "description": {
                                                            "type": "string"
                                                        },
                                                        "attachments": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "object"
                                                            }
                                                        },
                                                        "extention": {
                                                            "type": "string"
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "format": "datetime"
                                                        },
                                                        "created_by": {
                                                            "type": "string"
                                                        },
                                                        "deleted_at": {
                                                            "type": "string",
                                                            "format": "datetime"
                                                        },
                                                        "deleted_by": {
                                                            "type": "string"
                                                        },
                                                        "attachment": {
                                                            "type": "string"
                                                        },
                                                        "updated_at": {
                                                            "type": "string",
                                                            "format": "datetime"
                                                        },
                                                        "updated_by": {
                                                            "type": "string"
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/post/{id}": {
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Untuk mengubah post",
                "description": "",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id postingan",
                        "required": true,
                        "schema": {}
                    },
                    {
                        "name": "app-token",
                        "in": "query",
                        "description": "",
                        "required": false,
                        "schema": {}
                    }
                ],
                "requestBody": {
                    "description": "",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "description":{
                                        "type": "string"
                                    },
                                    "attachment":{
                                        "type": "array",
                                        "items":{
                                            "properties": {
                                                "extention":{
                                                    "type": "string",
                                                    "format": "photo"
                                                },
                                                "content":{
                                                    "type": "string"
                                                }
                                            }
                                        }

                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "type": "object",
                                                    "properties": {
                                                        "id": {
                                                            "type": "string"
                                                        },
                                                        "file": {
                                                            "type": "string",
                                                            "format": "file"
                                                        },
                                                        "contents": {
                                                            "type": "string"
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "untuk menghapus post",
                "description": "",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id postingan",
                        "required": true,
                        "schema": {}
                    },
                    {
                        "name": "app-token",
                        "in": "query",
                        "description": "",
                        "required": false,
                        "schema": {}
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "untuk melihat detail post",
                "description": "",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id postingan",
                        "required": true,
                        "schema": {}
                    },
                    {
                        "name": "app-token",
                        "in": "query",
                        "description": "",
                        "required": false,
                        "schema": {}
                    }
                ],
                "requestBody": {
                    "description": "",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "attachment": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "type": "object",
                                                    "properties": {
                                                        "id_postingan": {
                                                            "type": "string"
                                                        },
                                                        "file": {
                                                            "type": "string",
                                                            "format": "file"
                                                        },
                                                        "content": {
                                                            "type": "string"
                                                        },
                                                        "date": {
                                                            "type": "string",
                                                            "format": "date"
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/users/{id}/post": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Untuk melihat postingan postingan user",
                "description": "",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id user",
                        "required": true,
                        "schema": {}
                    },
                    {
                        "name": "app-token",
                        "in": "query",
                        "description": "",
                        "required": false,
                        "schema": {}
                    }
                ],
                "requestBody": {
                    "description": "",
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object",
                                                        "properties": {
                                                            "id": {
                                                                "type": "string"
                                                            },
                                                            "file": {
                                                                "type": "string",
                                                                "format": "file"
                                                            },
                                                            "contents": {
                                                                "type": "string"
                                                            },
                                                            "total_like": {
                                                                "type": "string"
                                                            },
                                                            "total_comment": {
                                                                "type": "string"
                                                            }
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}